build: pin the rust toolchain
CI resolved stable while the workstation had an older release, so clippy
ran different lint sets in the two places and a locally green tree failed on
push. rust-toolchain.toml is now the single source of the version and the
workflows install from it.

Also fixes the two lints this exposed: chunks_exact with a constant size becomes
as_chunks, and the statvfs fields need no widening on the targets we build.

+++ crates/is-pipeline/src/disk.rs
@@ -17,7 +17,9 @@ pub fn free_bytes(path: &Path) -> Option<u64> {
     if unsafe { libc::statvfs(c.as_ptr(), &mut st) } != 0 {
         return None;
     }
-    Some(st.f_bavail as u64 * st.f_frsize as u64)
+    // 这两个字段在我们支持的目标上都是 u64。若将来加 32 位目标，
+    // 这里需要显式加宽，否则大磁盘会溢出。
+    Some(st.f_bavail * st.f_frsize)
 }
 
 pub fn human_bytes(n: u64) -> String {
